feat: fix notification bug

+++ app/src/main/java/com/example/hanaparal/data/model/StudyGroup.kt
@@ -8,7 +8,7 @@ data class StudyGroup(
     val subject: String = "",
     val adminId: String = "",
     val adminName: String = "",
-    val maxMembers: Int = 10,
+    val maxMembers: Long = 10L,
     val createdAt: Timestamp? = null,
     val announcementHeader: String = ""
 )
